refactor(finding): move views + urls into dojo/finding/ui/ [finding Phase 5]
